Extrae views de crear_conjuntos_de_enunciados
