+++ src/enunciados/views/conjuntos_de_enunciados/conjuntos_de_enunciados_forms.py
@@ -4,8 +4,7 @@ from django.utils import timezone
 from django.utils.html import format_html
 from django.utils.safestring import mark_safe
 from django.utils.translation import gettext_lazy as _
-from enunciados.models import ConjuntoDeEnunciadosConCuatrimestre, Final, \
-    Parcial, Practica
+from enunciados.models import ConjuntoDeEnunciadosConCuatrimestre, Final, Parcial, Practica
 from enunciados.utils import conjuntos_url_parser
